Move some utils up to `common/`
The dapps will need these utils.
