+++ core/scripts/local/VoteGasEstimation.js
@@ -1,5 +1,5 @@
-const { RegistryRolesEnum, VotePhasesEnum } = require("../../utils/Enums.js");
-const { getRandomSignedInt, getRandomUnsignedInt } = require("../../utils/Random.js");
+const { RegistryRolesEnum, VotePhasesEnum } = require("../../../common/Enums.js");
+const { getRandomSignedInt, getRandomUnsignedInt } = require("../../../common/Random.js");
 const { moveToNextRound, moveToNextPhase } = require("../../utils/Voting.js");
 
 const Registry = artifacts.require("Registry");
