+++ core/utils/Voting.js
@@ -1,6 +1,6 @@
 const web3 = require("web3");
 
-const { VotePhasesEnum } = require("./Enums.js");
+const { VotePhasesEnum } = require("../../common/Enums.js");
 
 const secondsPerDay = web3.utils.toBN(86400);
